% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtime_dm.R
\docType{data}
\name{dm_model}
\alias{dm_model}
\title{Dirichlet-Multinomial Stan Model
Data:
S - number of signatures
N - number of categories in signature (e.g. 96 for SBS)
In stan, the N is implicit given the data vector Y
Y - mutation counts in epoch
alpha dirichlet prior for each signature weight
signatures - reference profiles for active signatures}
\format{
An object of class \code{character} of length 1.
}
\usage{
dm_model
}
\description{
Dirichlet-Multinomial Stan Model
Data:
S - number of signatures
N - number of categories in signature (e.g. 96 for SBS)
In stan, the N is implicit given the data vector Y
Y - mutation counts in epoch
alpha dirichlet prior for each signature weight
signatures - reference profiles for active signatures
}
\keyword{datasets}
