% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{helper_simulate_mutations}
\alias{helper_simulate_mutations}
\title{Apply matrix multiplication to get mutation counts from signatures}
\usage{
helper_simulate_mutations(sigs, counts, props = NULL, mut_classes = NULL)
}
\arguments{
\item{sigs}{Signature matrix}

\item{counts}{N mutations in (sub)sample}

\item{props}{proportions of each signature}

\item{mut_classes}{96-long character to name mutation classes}
}
\description{
Apply matrix multiplication to get mutation counts from signatures
}
