% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtime_hdp.R
\name{run_hdp_multichain}
\alias{run_hdp_multichain}
\title{Run HDP through parallel chains}
\usage{
run_hdp_multichain(model, dpindex, initcc = 2, burnin = 2500, n_chains = 4)
}
\description{
Run HDP through parallel chains
}
