% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtime_hdp.R
\name{sigtime_hdp}
\alias{sigtime_hdp}
\title{Run heirarchical dirichlet process for one or multiple samples}
\usage{
sigtime_hdp(data, priors, infer = FALSE)
}
\arguments{
\item{data}{A matrix of mx96 mutation counts, where m is clonal then subclonal for each sample.}

\item{priors}{A matrix of prior mutational signatures present in the whole dataset.}

\item{infer}{Attempt to infer novel signatures in the dataset.}
}
\value{
A HDP object
}
\description{
Run heirarchical dirichlet process for one or multiple samples
}
