% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate sample clonal-subclonal mutation profile}
\usage{
simulate(sigs, total, clonal_prop, cs_sig_props)
}
\arguments{
\item{sigs}{Signature matrix}

\item{total}{N mutations in sample}

\item{clonal_prop}{clonal proportion of total (numeric). Remainder of mutations will be subclonal}

\item{cs_sig_props}{proportion of signatures in clonal and subclonal (list)}
}
\description{
Simulate sample clonal-subclonal mutation profile
}
